requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-debug-toolbar==4.4.2
python-dotenv==1.0.1
Pillow==10.3.0
